+++ resource/styles/presentation.js
@@ -1,0 +1,1 @@
+import '../css/crowi-reveal.scss';