+++ CHANGES.md
@@ -4,8 +4,10 @@ CHANGES
 ## 1.0.0
 
 * Feature: Plugin mechanism
+* Feature: Switchable LineBreaks ON/OFF from admin page
 * Improvement: Exclude Environment-dependency
-* Support: abolish gulp
+* Improvement: Enhanced linker
+* Support: Add Dockerfile
+* Support: Abolish gulp
 * Support: LiveReload
-* Support: update libs
-
+* Support: Update libs